requirements.txt
langchain-core==0.3.21
langchain-mistralai==0.2.2
python-dotenv==1.0.1
psycopg2-binary==2.9.10
fastapi==0.115.5
uvicorn==0.32.1
Jinja2==3.1.4
python-multipart==0.0.18